feat: poc extracting ES initializer to its own module.
