+++ zeebe/exporters/elasticsearch-exporter/pom.xml
@@ -189,6 +189,18 @@
       <artifactId>zeebe-test-util</artifactId>
       <scope>test</scope>
     </dependency>
+    <dependency>
+      <groupId>io.camunda</groupId>
+      <artifactId>zeebe-elasticsearch-initializer</artifactId>
+      <version>8.6.10-SNAPSHOT</version>
+      <scope>compile</scope>
+    </dependency>
+    <dependency>
+      <groupId>io.camunda</groupId>
+      <artifactId>zeebe-elasticsearch-shared</artifactId>
+      <version>8.6.10-SNAPSHOT</version>
+      <scope>compile</scope>
+    </dependency>
 
   </dependencies>
