+++ zeebe/exporters/zeebe-elasticsearch-shared/src/main/java/io/camunda/zeebe/exporter/ElasticsearchExporterConfiguration.java
