+++ zeebe/exporters/zeebe-elasticsearch-shared/src/main/java/io/camunda/zeebe/exporter/RecordSequence.java
